+++ src/test/java/nextstep/subway/path/acceptance/PathAcceptanceTest.java
@@ -4,21 +4,20 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.line.acceptance.LineAcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-import static nextstep.subway.line.acceptance.LineAcceptanceTest.지하철_노선_등록되어_있음;
+import static nextstep.subway.line.acceptance.LineAcceptanceFixture.지하철_노선_등록되어_있음;
 import static nextstep.subway.line.acceptance.LineSectionAcceptanceTest.지하철_노선에_지하철역_등록_요청;
+import static nextstep.subway.path.acceptance.PathAcceptanceFixture.지하철_최단_경로_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
@@ -55,35 +54,29 @@ public class PathAcceptanceTest extends AcceptanceTest {
 
 
         LineRequest 신분당선_요청 = new LineRequest("신분당선", "bg-red-300", 강남역.getId(), 양재역.getId(), 10);
-        신분당선 = LineAcceptanceTest.지하철_노선_등록되어_있음(신분당선_요청).as(LineResponse.class);
+        신분당선 = 지하철_노선_등록되어_있음(신분당선_요청).as(LineResponse.class);
 
         LineRequest 이호선_요청 = new LineRequest("이호선", "bg-yellow-420", 교대역.getId(), 강남역.getId(), 10);
-        이호선 = LineAcceptanceTest.지하철_노선_등록되어_있음(이호선_요청).as(LineResponse.class);
+        이호선 = 지하철_노선_등록되어_있음(이호선_요청).as(LineResponse.class);
 
         LineRequest 삼호선_요청 = new LineRequest("삼호선", "bg-green-500", 교대역.getId(), 양재역.getId(), 5);
-        삼호선 = LineAcceptanceTest.지하철_노선_등록되어_있음(삼호선_요청).as(LineResponse.class);
+        삼호선 = 지하철_노선_등록되어_있음(삼호선_요청).as(LineResponse.class);
         지하철_노선에_지하철역_등록_요청(삼호선, 교대역, 남부터미널역, 3);
     }
 
     @DisplayName("최단 경로 조회한 경우")
     @Test
     void shortest_path() {
-
-        ExtractableResponse<Response> 지하철_경로_조회_응답 = RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .param("source", 교대역.getId())
-                .param("target", 양재역.getId())
-                .when().get("/paths")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> 지하철_경로_조회_응답 = 지하철_최단_경로_요청(교대역.getId(), 양재역.getId());
 
         assertAll(
                 () -> assertThat(지하철_경로_조회_응답.statusCode()).isEqualTo(HttpStatus.OK.value()),
-                () -> assertThat(지하철_경로_조회_응답.as(PathResponse.class).getDistance()).isEqualTo(5)
-
+                () -> assertThat(지하철_경로_조회_응답.as(PathResponse.class).getDistance()).isEqualTo(5),
+                () -> assertThat(지하철_경로_조회_응답.as(PathResponse.class).getStations().size()).isEqualTo(3)
         );
     }
+
+
 
     @DisplayName("출발역과 도착역이 같은 경우 경로 조회 불가")
     @Test
